pyproject.toml
[tool.poetry]
name = "django-chatbot"
version = "0.1.0"
description = ""
authors = ["UilSiqueira <uildsiqueira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=4.2"
djangorestframework = "^3.16.0"
celery = "^5.5.3"
redis = "^6.2.0"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
django-redis = ">=5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
